Abstract access presentation into a utility
